requirements.txt
pandas
numpy
yfinance
fredapi
python-dotenv
matplotlib
streamlit
